feat: mermaid output support
